Autotest.sh ignore tiny numerical error
In some certain cases, abacus failes to output result. This should be treated as failing to pass the test. 
However, in most common cases, the result error is slightly above threshold. This might be errors introduced by various of math libs. Currently we focus on a proper developing pipeline, and later deciding which result is canonical. 
Now, Autotest.sh will treat the former cases as 'fatal error', and exit with code 1. For the latter cases, the test script will raise a failed prompt, but will exit normally.